Add detailed README and Docker build workflow
An elaborate README.md document was added, which provides detailed information on the Docker images contained in the repo, including their usages and naming conventions. Also, a Docker build workflow was created for GitHub actions, which will automatically build and push images to Docker Hub upon changes to the main branch. A Dockerfile for CentOS 7 was also added to initialize GetPageSpeed LTS repository, keeping the system updated.