Implement dynamic Docker build strategy
A new strategy has been implemented for the Docker build using a matrix from a JSON file. This approach allows flexibility in defining the Docker image builds and pushes, as it moves their configuration into the separate JSON file. The configuration is read, set as an output for a new "config" job, and then used in the original build-and-push job.