Update Docker GitHub workflow configuration
The configurations in the docker-build.yml file in GitHub workflows have been updated to separate the setup and configuration processes into different jobs. Previously encapsulated in "build-and-push", the configuration has been moved to a new "configure" job for better clarity and separation of concerns. The "matrix-build-and-push" job now depends on this new "configure" job.