Update Dockerfile maintainer labels and tag names
All Dockerfiles were updated to change the maintainer email address to the correct info email for each respective version. The docker-matrix.json file was updated too to adjust the names and tags associated with each version in the Dockerfiles. Now they align to the simplified 'el7', 'el8', and 'el9' tags, rather than the previous '-lts' suffix.